+++ test/index.js
@@ -27,6 +27,27 @@ test('Receive all expected events from a successful job', t => {
   })
 })
 
+test('Receive all expected events from a delayed successful job', t => {
+  t.plan(9)
+  const job = queue.enqueue('succeed', {id: 'succeed', delay: 1000})
+  job
+  .on('start', j => {
+    t.equal(j.status, 'start', 'Correct status')
+    t.equal(job.id, j.id, 'ID passed')
+    t.pass('Start event handled')
+  })
+  .on('progress', j => {
+    t.equal(j.status, 'progress', 'Correct status')
+    t.equal(job.id, j.id, 'ID passed')
+    t.pass('Progress event handled')
+  })
+  .on('finish', j => {
+    t.equal(j.status, 'finish', 'Correct status')
+    t.equal(job.id, j.id, 'ID passed')
+    t.pass('Finish event handled')
+  })
+})
+
 test('Receive expected events from a failed job', t => {
   t.plan(3)
   const job = queue.enqueue('fail', {id: 'fail'})