+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+* Delayed job option
+
 ## [0.1.2] - 2016-09-26
 ### Fixed
 * No longer throwing exception when shutting down before a job has run